The 2 changes in bzflag.cxx were to handle (a) a "-psn" flag that MacOS X sends to all double-clicked apps, and (b) looking in the right place (inside the app bundle, not in /usr/local) for the data files.
On the Mac-specific side, this commit adds a Project Builder directory
(src/platform/MacBZFlag.pbproj) inside of which are text files that
direct Project Builder on how to build the app. Project Builder is
based on "Jam", a command-line tool similar to "Make".

Usually the file "InfoPlist.strings" lives under a directory called
"English.lproj", but in the interests of keeping the Mac stuff under
control and naming all Mac stuff with the prefix "Mac", I moved it to
the MacBZFlag.pbproj directory instead, which is very unusual for a
Project Builder project (thus this note in case someone wonders what
happened).

MacBZFlag-prefix.h just helps with #includes a bit. i hope to make
this a precompiled header later to speed up builds quite a bit.

MacBZFlag.icns is a MacOS X binary icon file. Hope it's handled
correctly. Easy to fix if there's a problem.

MacDisplay now uses Carbon Events instead of the old
WaitNextEvent. This allows Mac users to use multi-button mice with as
many buttons as they like. This is what first got me into working on
BZFlag. I got a new Logitech MX700 and couldn't use anything but the
two main buttons. Now I have *8* working buttons on my mouse. Woohoo!

MacWindow.cxx no longer hides the mouse since that makes it really
hard to drive.

(As an aside, is one big checkin like this with lots of comments the
usual way to work, or do you guys prefer separate checkins with
comments for each file?)


git-svn-id: https://bzflag.svn.sourceforge.net/svnroot/bzflag/branches/v1_7@1099 08b3d480-bf2c-0410-a26f-811ee3361c24