apply crazy tuponies patch that works around a recent bug in the SDL library's audio layer (sf patch 1546333). to paraphrase him, he found a bug inside SDL's audio processing that causes a segfault when resampling audio (typically alsa+dmix).  even though the bug is inside SDL, this provides a workaround.
also he notes, as a note to yon devs, that audio files are resampled down to 11025 or 22050 so there's no need to have files sampled at 48000.  he notes that he doesn't know if the resampling is good on all platforms (needs testing), but there are a few that aren't at 22050 that should probably be resampled down.


git-svn-id: https://bzflag.svn.sourceforge.net/svnroot/bzflag/trunk@13275 08b3d480-bf2c-0410-a26f-811ee3361c24