Do not send plugin messages from the server through the message filter
Server messages generate a server message event not a raw chat event.
Plugin generated events that come from the server should not generate
raw chat events which get filtered since that generates two events
for the message - one for the raw chat event and another for the server
message event.

git-svn-id: https://bzflag.svn.sourceforge.net/svnroot/bzflag/trunk@19534 08b3d480-bf2c-0410-a26f-811ee3361c24